requirements.txt
fastapi
boto3
uvicorn[standard]
python-multipart
python-dotenv